+++ services/scf/CHANGELOG.md
@@ -1,2 +1,6 @@
+## v0.2.0
+- **Feature:** Add field `OrgId` in model `OrgManager`
+- **Feature:** Add new model `OrganizationCreateBffResponse` and `SpaceCreatedBffResponse`
+
 ## v0.1.0
 - **New:** STACKIT Cloud Foundry module